Blender Animation
So, despite the fact that the two test animations are 90 frames instead
of 32, the build scripts (on my machine) take the *same* time to
complete. It's safe to say I'm probably gonna keep using Blender 2.8X
for this as it Just Works™ and hasn't hung or crashed once. Decided
against rendering at 256px because in the instance I need even larger
cursors for a TV or some shit, I found there's literally an addon called
"Copy Render Settings" in blender, so for now I'll render at the usual
128 since changing basic settings across all scenes is about two clicks.
Only possible complaint I have with the workflow is the addon can't copy
certain niche settings like compositor performance settings.